chore: use correct icon and remove unused
