+++ assets/styles/customizer.scss
@@ -271,10 +271,6 @@ button.add-to-cart {
         fill: var(--accents-7);
       }
     }
-
-    .options-name {
-      height: calc(100% - 150px);
-    }
   }
 
   > .list-view {
