+++ components/Customizer/Block.tsx
@@ -49,7 +49,7 @@ export const Block = ({
   return (
     <div ref={blockRef} id={subs.categoryName.toLowerCase()} className='flex flex-col w-full select-none'>
       <div
-        className={classNames('flex items-center justify-between w-full', incompatibleCats?.some(
+        className={classNames('flex items-center justify-between w-full h-20', incompatibleCats?.some(
           (cat: any) =>
             cat ===
             subs.categoryName
