+++ components/icons/Info.tsx
@@ -1,22 +1,17 @@
-const Info = ({ ...props }) => {
-  return (
-    <svg
-      viewBox="0 0 24 24"
-      width="24"
-      height="24"
-      stroke="currentColor"
-      strokeWidth="1.5"
-      strokeLinecap="round"
-      strokeLinejoin="round"
-      fill="none"
-      shapeRendering="geometricPrecision"
-      {...props}
-    >
-      <circle cx="12" cy="12" r="10" fill="transparent" />
-      <path d="M12 8v4" stroke="currentColor" />
-      <path d="M12 16h.01" stroke="currentColor" />
-    </svg>
-  )
-}
 
-export default Info
+const InfoIcon = ({ ...props }) => (
+  <svg
+    width="18"
+    height="18" 
+    viewBox="0 0 18 18" 
+    fill="none" 
+    xmlns="http://www.w3.org/2000/svg" 
+    {...props}
+  >
+    <path d="M9 0.0626831L8.74679 0.0662719C3.91182 0.206173 0.0625 4.16186 0.0625 9.00018L0.0660888 9.25339C0.205989 14.0884 4.16167 17.9377 9 17.9377L9.25321 17.9341C14.0882 17.7942 17.9375 13.8385 17.9375 9.00018L17.9339 8.74698C17.794 3.91201 13.8383 0.0626831 9 0.0626831ZM8.77663 1.44091L9 1.43768C13.0939 1.43768 16.4411 4.69487 16.5593 8.77681L16.5626 9.01012C16.5626 13.0941 13.3053 16.4413 9.22337 16.5595L8.99006 16.5628C4.90609 16.5628 1.5589 13.3055 1.44073 9.22356L1.43743 8.99025C1.43743 4.90627 4.69468 1.55908 8.77663 1.44091Z" fill="currentColor"/>
+    <path d="M9.00065 8.31268C9.34871 8.31268 9.63635 8.57132 9.68188 8.90689L9.68815 9.00018V13.5835C9.68815 13.9632 9.38035 14.271 9.00065 14.271C8.6526 14.271 8.36495 14.0124 8.31943 13.6768L8.31315 13.5835V9.68768H8.08398C7.73593 9.68768 7.44828 9.42904 7.40276 9.09347L7.39648 9.00018C7.39648 8.65213 7.65513 8.36448 7.99069 8.31896L8.08398 8.31268H9.00065Z" fill="currentColor"/>
+    <path d="M8.76927 4.64569L8.87624 4.65185C9.33244 4.70474 9.68685 5.09179 9.68685 5.56236C9.68685 6.06856 9.27638 6.47902 8.77018 6.47902C8.26399 6.47902 7.85352 6.06855 7.85354 5.56734C7.8501 5.09257 8.20861 4.70118 8.66912 4.6511L8.76927 4.64569Z" fill="currentColor"/>
+  </svg>
+);
+
+export default InfoIcon;
